incremental cleanup for x-platform
