+++ make.bat
@@ -7,7 +7,7 @@ mkdir out
 if "%1"=="debug" (
     echo Compiling debug version...
     mkdir "out\debug\"
-    cl /EHsc src\main.cpp src\nayuki-qr\qrcodegen.cpp src\webserver.cpp src\dataloader.cpp src\qr_to_bmp.cpp src\window.cpp src\gethost.cpp src\randompath.cpp src/settings/settings.cpp src\icon\icon.res /Fe:out\debug\Axon.exe /Fo:out\debug\ /I "%VCPKG_ROOT%\%TRIPLET%\include" /link /LIBPATH:"%VCPKG_ROOT%\%TRIPLET%\lib"
+    cl /EHsc src\main.cpp src\common.cpp src\nayuki-qr\qrcodegen.cpp src\webserver.cpp src\dataloader.cpp src\qr_to_bmp.cpp src\window.cpp src\gethost.cpp src/settings/settings.cpp src\icon\icon.res /Fe:out\debug\Axon.exe /Fo:out\debug\ /I "%VCPKG_ROOT%\%TRIPLET%\include" /link /LIBPATH:"%VCPKG_ROOT%\%TRIPLET%\lib"
     xcopy src\static out\debug\static /E /I /H /Y
     xcopy %VCPKG_ROOT%\%TRIPLET%\bin\* out\debug\* /Y
 ) else (
