+++ src/main.cpp
@@ -1,5 +1,3 @@
-#include <winsock2.h>
-#include <windows.h>
 #include <iostream>
 #include <string>
 #include <thread>
@@ -10,31 +8,11 @@
 #include "dataloader.hpp"
 #include "qr_to_bmp.hpp"
 #include "window.hpp"
-#include "randompath.hpp"
+#include "common.hpp"
 #include "settings/settings.hpp"
-
-#pragma comment(lib, "ws2_32.lib")
-#pragma comment(lib, "iphlpapi.lib")
 
 char QR_1_STR[2] = {(char)219, (char)219};
 char QR_0_STR[2] = {(char)32, (char)32};
-
-#define QR_FALSE
-
-std::string getParentDirectory(const std::string& fullPath) {
-    size_t lastSlashPos = fullPath.find_last_of("\\/");
-    if (lastSlashPos != std::string::npos) {
-        return fullPath.substr(0, lastSlashPos + 1); // Include trailing slash
-    }
-    return "";
-}
-
-std::string getExecutableDirectory() {
-    char buffer[MAX_PATH];
-    GetModuleFileNameA(nullptr, buffer, MAX_PATH);
-    std::string fullPath(buffer);
-    return fullPath.substr(0, fullPath.find_last_of("\\/"));
-}
 
 std::string ROOT_DIR;
 AXONSETTINGSCONF SETTINGS;
@@ -64,7 +42,7 @@ int main(int argc, char* argv[]) {
     std::string data = loadFile(fname);
 
     // Build url
-    std::string url_path = '/' + generatePath();
+    std::string url_path = '/' + randAN(5);
     std::string url = "http://" + SETTINGS.host + ":" + std::to_string(SETTINGS.port) + url_path;
     
     // Start webserver
