--- src/randompath.hpp
@@ -1,1 +1,0 @@
-std::string generatePath();
